+++ genesis-volatility/adapter.js
@@ -1,6 +1,5 @@
 const { Requester, Validator } = require('@chainlink/external-adapter')
-
-const API_KEY = process.env.API_KEY
+const { util } = require('@chainlink/ea-bootstrap')
 
 const customParams = {
   symbol: ['base', 'from', 'coin', 'symbol'],
@@ -8,6 +7,8 @@ const customParams = {
 }
 
 const execute = (input, callback) => {
+  const API_KEY = util.getRandomRequiredEnv('API_KEY')
+
   const validator = new Validator(input, customParams)
   if (validator.error) return callback(validator.error.statusCode, validator.errored)